+++ src/components/data-table/utils/search.ts
@@ -1,0 +1,21 @@
+/**
+ * Preprocess search terms before sending to API
+ */
+export function preprocessSearch(searchTerm: string): string {
+  if (!searchTerm) return "";
+
+  // Trim whitespace
+  let processed = searchTerm.trim();
+
+  // Remove excessive whitespace within the search term
+  processed = processed.replace(/\s+/g, " ");
+
+  // Check for minimum length after processing (e.g., if it's just spaces)
+  if (processed.length < 1) return "";
+
+  // Basic sanitization for API safety
+  // Remove any potentially harmful characters for the backend
+  processed = processed.replace(/[<>]/g, "");
+
+  return processed;
+}
